requirements.txt
langchain
openai
huggingface_hub
streamlit
python-dotenv
langchain_community